Remove some redundant classes
