+++ src/SFA.DAS.EmployerAccounts/Models/EmployerAgreement/RemoveOrganisationView.cs
@@ -13,15 +13,4 @@ namespace SFA.DAS.EmployerAccounts.Models.EmployerAgreement
         public OrganisationType LegalEntitySource { get; set; }
         public bool HasSignedAgreement { get; set; }
     }
-
-
-    public class EmployerAgreementRemoved
-    {
-        public long EmployerAgreementId { get; set; }
-        public long LegalEntityId { get; set; } 
-        public string LegalEntityName { get; set; }
-        public long AccountLegalEntityId { get; set; }
-        public EmployerAgreementStatus PreviousStatus { get; set; }
-        public bool Signed { get; set; }
-    }
 }
