--- src/SFA.DAS.EmployerAccounts/Queries/GetAccountEmployerAgreementsRemove/GetAccountEmployerAgreementsRemoveRequest.cs
@@ -1,10 +1,0 @@
-﻿using MediatR;
-
-namespace SFA.DAS.EmployerAccounts.Queries.GetAccountEmployerAgreementsRemove
-{
-    public class GetAccountEmployerAgreementsRemoveRequest : IAsyncRequest<GetAccountEmployerAgreementsRemoveResponse>
-    {
-        public string HashedAccountId { get; set; }
-        public string UserId { get; set; }
-    }
-}
