--- src/SFA.DAS.EmployerAccounts/Queries/GetAccountEmployerAgreementsRemove/GetAccountEmployerAgreementsRemoveResponse.cs
@@ -1,10 +1,0 @@
-﻿using System.Collections.Generic;
-using SFA.DAS.EmployerAccounts.Models.EmployerAgreement;
-
-namespace SFA.DAS.EmployerAccounts.Queries.GetAccountEmployerAgreementsRemove
-{
-    public class GetAccountEmployerAgreementsRemoveResponse
-    {
-        public List<RemoveEmployerAgreementView> Agreements { get; set; }
-    }
-}
