--- src/SFA.DAS.EmployerAccounts/Queries/GetAccountEmployerAgreementsRemove/GetAccountEmployerAgreementsRemoveValidator.cs
@@ -1,52 +1,0 @@
-using System;
-using System.Threading.Tasks;
-using SFA.DAS.Authorization;
-using SFA.DAS.EmployerAccounts.Data;
-using SFA.DAS.Validation;
-
-namespace SFA.DAS.EmployerAccounts.Queries.GetAccountEmployerAgreementsRemove
-{
-    public class GetAccountEmployerAgreementsRemoveValidator : IValidator<GetAccountEmployerAgreementsRemoveRequest>
-    {
-        private readonly IMembershipRepository _membershipRepository;
-
-        public GetAccountEmployerAgreementsRemoveValidator(IMembershipRepository membershipRepository)
-        {
-            _membershipRepository = membershipRepository;
-        }
-
-        public ValidationResult Validate(GetAccountEmployerAgreementsRemoveRequest item)
-        {
-            throw new NotImplementedException();
-        }
-
-        public async Task<ValidationResult> ValidateAsync(GetAccountEmployerAgreementsRemoveRequest item)
-        {
-            var validationResult = new ValidationResult();
-
-            if (string.IsNullOrEmpty(item.HashedAccountId))
-            {
-                validationResult.AddError(nameof(item.HashedAccountId));
-            }
-
-            if (string.IsNullOrEmpty(item.UserId))
-            {
-                validationResult.AddError(nameof(item.UserId));
-            }
-
-            if (!validationResult.IsValid())
-            {
-                return validationResult;
-            }
-
-            var member = await _membershipRepository.GetCaller(item.HashedAccountId, item.UserId);
-
-            if (member == null || member.RoleId != (short)Role.Owner)
-            {
-                validationResult.IsUnauthorized = true;
-            }
-
-            return validationResult;
-        }
-    }
-}
